warning
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@4869 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/r0drv/solaris/memobj-r0drv-solaris.c
@@ -83,7 +83,7 @@ static uint64_t rtR0MemObjSolarisVirtToPhys(struct hat* hatSpace, caddr_t virtAd
 #if 0
     uint64_t physAddr = PAGE_SIZE * pfn;
 #else
-    uint64_t physAddr = ((uint64_t)pfn << MMU_PAGESHIFT) | ((uint64_t)virtAddr & MMU_PAGEOFFSET);
+    uint64_t physAddr = ((uint64_t)pfn << MMU_PAGESHIFT) | ((uintptr_t)virtAddr & MMU_PAGEOFFSET);
 #endif
     return physAddr;
 }
